Sound Explorer and more audio files
Began work on the “Sound Explorer” for each tile in the Sound Palette. Wben you click the magnifying glass on a tile in the Sound Palette the generator screen disappears and a list of every single sound will appear with the ablity to test that sound and to select it as the active sound.

Plus even more audio files!