Removed virtual keyboard
And now sounds can only be added by clicking a button. Currently there is one on each Sound Palette Tile in the Sound Palette. Next there will be a button for each tile in Sound Explore mode.

This is all part of making the site reponsive down to smart phone.

~Gray